refactor: improve global persistent subscription
