+++ lib/app/features/feed/notifications/providers/notifications/like_notification_handler.c.dart
@@ -8,7 +8,7 @@ import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
 import 'package:ion/app/features/feed/data/models/entities/reaction_data.c.dart';
 import 'package:ion/app/features/feed/notifications/data/repository/likes_repository.c.dart';
 import 'package:ion/app/features/ion_connect/ion_connect.dart';
-import 'package:ion/app/features/ion_connect/providers/ion_connect_persistent_subscription.c.dart';
+import 'package:ion/app/features/ion_connect/model/persistent_subscription_event_handler.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'like_notification_handler.c.g.dart';
