+++ lib/app/features/feed/notifications/providers/notifications/quote_notification_handler.c.dart
@@ -7,7 +7,7 @@ import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
 import 'package:ion/app/features/feed/data/models/entities/modifiable_post_data.c.dart';
 import 'package:ion/app/features/feed/notifications/data/repository/comments_repository.c.dart';
 import 'package:ion/app/features/ion_connect/ion_connect.dart';
-import 'package:ion/app/features/ion_connect/providers/ion_connect_persistent_subscription.c.dart';
+import 'package:ion/app/features/ion_connect/model/persistent_subscription_event_handler.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'quote_notification_handler.c.g.dart';
