--- lib/app/features/feed/notifications/providers/notifications_subscription_provider.c.dart
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'notifications_subscription_provider.c.g.dart';
-
-@riverpod
-void notificationsSubscription(Ref ref) {}
