+++ lib/app/features/ion_connect/model/persistent_subscription_event_handler.dart
@@ -1,0 +1,7 @@
+import 'package:ion/app/features/ion_connect/ion_connect.dart';
+
+abstract class PersistentSubscriptionEventHandler {
+  bool canHandle(EventMessage eventMessage);
+
+  Future<void> handle(EventMessage eventMessage);
+}
